configure Irepository for db

+++ Transacciones.Core/Entities/Account/Accounts.cs
@@ -4,7 +4,7 @@
     {
         public required Guid Id { get; set; }
         public required string AccountNumber { get; set; }
-        public required double Balance { get; set; }
+        public required decimal Balance { get; set; }
         public required string Holder { get; set; }
         public required DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public required bool IsActive { get; set; } = true;
